can take the place of master

+++ controllers/user_controller.py
@@ -1,4 +1,6 @@
 from fastapi import HTTPException
+from models.shipper import ShipperStatus
+from utils.roles import Role
 from models.order_delivery import DeliveryStatusEnum
 from models.payment import Payment, PaymentStatus
 from services.payment_service import insert_payment_to_db
@@ -11,7 +13,7 @@ from services.shipper_service import delete_shipper_by_id, get_shipper_by_id, up
 from schemas.shipper_schema import ShipperAssignedOrderDeliverySchema, ShipperSchema
 from services.user_service import delete_user_in_db_by_id, find_customer_by_id, is_me, update_user_in_db_by_id
 from services.menu_service import get_all_menu_items, get_menu_item_by_id
-from schemas.user_schema import CreateOrderSchema, CustomerResponseSchema
+from schemas.user_schema import CreateOrderSchema, CustomerResponseSchema, UserSchema
 
 # Get all menu items
 async def get_menu_items() -> list:
@@ -20,22 +22,23 @@ async def get_menu_items() -> list:
 
 
 # Get customer profile
-async def get_customer_profile(customer_id: str) -> dict:
+async def get_customer_profile(customer_id: str, current_user: UserSchema) -> dict:
     customer_id = customer_id.lower()
+    if not is_me(customer_id, current_user):
+        raise HTTPException(status_code=403, detail="You do not have access to this resource")
     
-        
-    
-    customer = await find_customer_by_id(customer_id=customer_id)
+    customer = await find_customer_by_id(current_user.id)
     return customer
 
 # Update customer profile
-async def update_customer_profile(customer: CustomerResponseSchema) -> dict:
+async def update_customer_profile(customer: CustomerResponseSchema, current_user: UserSchema) -> dict:
     customer_id = customer.customer_id
     
     # Can invoke api with "me" as customer_id
     # Need to check if passed_id is not really "me"
-    
-        
+    if not is_me(customer_id, current_user):
+        raise HTTPException(status_code=403, detail="You do not have access to this resource")
+    customer_id = current_user.id
     
     modified_count = await update_user_in_db_by_id(customer_id, customer.model_dump())
     if modified_count == 0:
@@ -44,10 +47,14 @@ async def update_customer_profile(customer: CustomerResponseSchema) -> dict:
     return {"message": f"Your account has been updated successfully"}
 
 # Delete customer profile
-async def delete_user_profile(customer_id: str) -> dict:
+async def delete_user_profile(customer_id: str, current_user: UserSchema) -> dict:
     customer_id = customer_id.lower()
     
-        
+    # Can invoke api with "me" as customer_id
+    # Need to check if passed_id is not really "me"
+    if not is_me(customer_id, current_user):
+        raise HTTPException(status_code=403, detail="You do not have access to this resource")
+    customer_id = current_user.id
     
     deleted_count = await delete_user_in_db_by_id(customer_id)
     if deleted_count == 0:
@@ -56,7 +63,7 @@ async def delete_user_profile(customer_id: str) -> dict:
     return {"message": f"Your account has been deleted successfully"}
 
 # Create order
-async def create_order(customer_id: str, order: CreateOrderSchema) -> dict:  # Order
+async def create_order(customer_id: str, order: CreateOrderSchema, current_user: UserSchema) -> dict:  # Order
     """
     There's two main tasks to be done here:
     1. Handle order creation (voucher has to be checked)
@@ -65,6 +72,13 @@ async def create_order(customer_id: str, order: CreateOrderSchema) -> dict:  # O
     
     """1. Handle order creation (voucher has to be checked)"""
     customer_id = customer_id.lower()
+    
+    # Can invoke api with "me" as customer_id
+    # Need to check if passed_id is not really "me"
+    if not is_me(customer_id, current_user):
+        raise HTTPException(status_code=403, detail="You do not have access to this resource")
+    customer_id = current_user.id
+    
     # Check if the customer exists
     if not await find_customer_by_id(customer_id):
         raise HTTPException(status_code=404, detail="Customer not found")
@@ -100,7 +114,13 @@ async def create_order(customer_id: str, order: CreateOrderSchema) -> dict:  # O
     return inserted_order  # result of task 1
 
 # Update password
-async def update_password(customer_id: str, old_password: str, new_password: str) -> dict:
+async def update_password(customer_id: str, old_password: str, new_password: str, current_user: UserSchema) -> dict:
+    # Can invoke api with "me" as customer_id
+    # Need to check if passed_id is not really "me"
+    if not is_me(customer_id, current_user):
+        raise HTTPException(status_code=403, detail="You do not have access to this resource")
+    customer_id = current_user.id
+    
     user = await find_customer_by_id(customer_id)
     if not user:
         raise HTTPException(status_code=404, detail="Customer not found")
@@ -115,7 +135,13 @@ async def update_password(customer_id: str, old_password: str, new_password: str
     return {"message": f"Password updated successfully"}
 
 # Update password shipper
-async def update_password_shipper(shipper_id: str, old_password: str, new_password: str) -> dict:
+async def update_password_shipper(shipper_id: str, old_password: str, new_password: str, current_user: UserSchema) -> dict:
+    # Can invoke api with "me" as shipper_id
+    # Need to check if passed_id is not really "me"
+    if not is_me(shipper_id, current_user):
+        raise HTTPException(status_code=403, detail="You do not have access to this resource")
+    shipper_id = current_user.id
+    
     shipper = await get_shipper_by_id(shipper_id)
     if not shipper:
         raise HTTPException(status_code=404, detail="Shipper not found")
@@ -131,7 +157,13 @@ async def update_password_shipper(shipper_id: str, old_password: str, new_passwo
 
 
 # Get shipper history
-async def get_shipper_history(shipper_id: str) -> list[dict]:
+async def get_shipper_history(shipper_id: str, current_user: UserSchema) -> list[dict]:
+    # Can invoke api with "me" as shipper_id
+    # Need to check if passed_id is not really "me"
+    if not is_me(shipper_id, current_user):
+        raise HTTPException(status_code=403, detail="You do not have access to this resource")
+    shipper_id = current_user.id
+    
     history = await read_delivery_history_by_shipper_id(shipper_id)
     if not history:
         raise HTTPException(status_code=404, detail="No history found")
@@ -140,7 +172,13 @@ async def get_shipper_history(shipper_id: str) -> list[dict]:
 
 
 # Get delivering orders
-async def read_assigned_order_delivery(shipper_id: str) -> dict:   # -> ShipperAssingedOrderDeliverySchema
+async def read_assigned_order_delivery(shipper_id: str, current_user: UserSchema) -> dict:   # -> ShipperAssingedOrderDeliverySchema
+    # Can invoke api with "me" as shipper_id
+    # Need to check if passed_id is not really "me"
+    if not is_me(shipper_id, current_user):
+        raise HTTPException(status_code=403, detail="You do not have access to this resource")
+    shipper_id = current_user.id
+    
     # Get all order delivery of shipper by shipper id
     all_order_delivery_of_shipper = await get_order_delivery_by_shipper_id(shipper_id)
     
@@ -177,8 +215,12 @@ async def read_assigned_order_delivery(shipper_id: str) -> dict:   # -> ShipperA
     ).model_dump()
 
 # Get delivery fee with respect to address
-async def read_delivery_fee(customer_id: str) -> dict:
+async def read_delivery_fee(customer_id: str, current_user: UserSchema) -> dict:
     customer_id = customer_id.lower()
+    if not is_me(customer_id, current_user):
+        raise HTTPException(status_code=403, detail="You do not have access to this resource")
+    customer_id = current_user.id
+    
     customer = await find_customer_by_id(customer_id)
     if not customer:
         raise HTTPException(status_code=404, detail="Customer not found")
@@ -196,7 +238,13 @@ async def read_delivery_fee(customer_id: str) -> dict:
     return {"delivery_fee": DeliveryFeeDict.delivery_fee_dict.get(address_to_calculate)}
 
 # Read orders history of a customer
-async def read_orders_history(customer_id: str) -> list[dict]:
+async def read_orders_history(customer_id: str, current_user: UserSchema) -> list[dict]:
+    # Can invoke api with "me" as customer_id
+    # Need to check if passed_id is not really "me"
+    if not is_me(customer_id, current_user):
+        raise HTTPException(status_code=403, detail="You do not have access to this resource")
+    customer_id = current_user.id
+    
     history = await read_order_history_by_customer_id(customer_id)
     if not history:
         raise HTTPException(status_code=404, detail="You have not made any orders")
@@ -204,13 +252,14 @@ async def read_orders_history(customer_id: str) -> list[dict]:
     return history
 
 # Get shipper profile
-async def get_shipper_profile(shipper_id: str) -> dict:
+async def get_shipper_profile(shipper_id: str, current_user: UserSchema) -> dict:
     shipper_id = shipper_id.lower()
     
-    # Can invoke api with "me" as shipper_id
+    # Can invoke api with "me" as customer_id
     # Need to check if passed_id is not really "me"
-    
-        
+    if not is_me(shipper_id, current_user):
+        raise HTTPException(status_code=403, detail="You do not have access to this resource")
+    shipper_id = current_user.id
     
     # Get shipper infor by shipper id
     shipper = await get_shipper_by_id(shipper_id)
@@ -221,13 +270,14 @@ async def get_shipper_profile(shipper_id: str) -> dict:
     return shipper
 
 # Update shipper profile
-async def update_shipper_profile(shipper: ShipperSchema) -> dict:
+async def update_shipper_profile(shipper: ShipperSchema, current_user: UserSchema) -> dict:
     shipper_id = shipper.shipper_id
     
     # Can invoke api with "me" as shipper_id
     # Need to check if passed_id is not really "me"
-    
-        
+    if not is_me(shipper_id, current_user):
+        raise HTTPException(status_code=403, detail="You do not have access to this resource")
+    shipper_id = current_user.id
     
     # Update shipper infor by shipper id
     modified_count = await update_shipper_by_id(shipper_id, shipper.model_dump())
@@ -237,10 +287,14 @@ async def update_shipper_profile(shipper: ShipperSchema) -> dict:
     return {"message": f"Your account has been updated successfully"}
 
 # Delete shipper profile
-async def delete_shipper_profile(shipper_id: str) -> dict:
+async def delete_shipper_profile(shipper_id: str, current_user: UserSchema) -> dict:
     shipper_id = shipper_id.lower()
     
-            
+    # Can invoke api with "me" as shipper_id
+    # Need to check if passed_id is not really "me"
+    if not is_me(shipper_id, current_user):
+        raise HTTPException(status_code=403, detail="You do not have access to this resource")
+    shipper_id = current_user.id
     
     deleted_count = await delete_shipper_by_id(shipper_id)
     if deleted_count == 0:
@@ -266,3 +320,11 @@ async def read_shipper_infor_by_order_id(order_id: str) -> dict:
     return {"shipper_name": shipper_who_shipped["name"],
             "address":shipper_who_shipped["address"],
             "phone_number":shipper_who_shipped["phone_number"]}
+
+
+# Logout
+async def logout(current_user: UserSchema) -> dict:
+    if current_user.role == Role.SHIPPER:
+        await update_shipper_by_id(current_user.id, {"account_status": ShipperStatus.INACTIVE})
+        
+    return {"message": "You have been logged out"}
